NTFS: Fix directory corruption on file delete.
NTFS appears to have invariant that 'End' entries in a node must have children
unless we're in a node with only leaf entries.  This was not previously
honoured.  So, honour that, and take advantage of it to simplify entry removal,
since it makes it trivial to find the largest leaf of any node.